Check autorizations

+++ LetsGame.Web/Services/GroupService.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using LetsGame.Web.Data;
 using LetsGame.Web.Helpers;
 using LetsGame.Web.Services.Igdb;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 
 namespace LetsGame.Web.Services
@@ -13,37 +14,42 @@ namespace LetsGame.Web.Services
         private readonly ApplicationDbContext _db;
         private readonly SlugGenerator _slugGenerator;
         private readonly IgdbClient _igdbClient;
+        private readonly ICurrentUserAccessor _currentUserAccessor;
+        private readonly UserManager<AppUser> _userManager;
 
-        public GroupService(ApplicationDbContext db, SlugGenerator slugGenerator, IgdbClient igdbClient)
+        public GroupService(
+            ApplicationDbContext db, 
+            SlugGenerator slugGenerator, 
+            IgdbClient igdbClient, 
+            ICurrentUserAccessor currentUserAccessor, 
+            UserManager<AppUser> userManager)
         {
             _db = db;
             _slugGenerator = slugGenerator;
             _igdbClient = igdbClient;
+            _currentUserAccessor = currentUserAccessor;
+            _userManager = userManager;
         }
 
         public Task<Group> FindBySlugAsync(string slug)
         {
-            return _db.Groups.FirstOrDefaultAsync(x => x.Slug == slug);
+            return _db.Groups.FirstOrDefaultAsync(x => x.Slug == slug &&
+                                                       x.Memberships.Any(x => x.UserId == CurrentUserId));
         }
 
-        public Task<string> GetSlugFromGroupNameAsync(string name)
-        {
-            return _slugGenerator.GenerateWithCheck(name, slug => _db.Groups.AnyAsync(x => x.Slug == slug));
-        }
-
-        public async Task<Group> CreateGroupAsync(string groupName, string ownerId, string ownerDisplayName)
+        public async Task<Group> CreateGroupAsync(string groupName, string ownerDisplayName)
         {
             var group = new Group
             {
                 Name = groupName,
-                Slug = await GetSlugFromGroupNameAsync(groupName)
+                Slug = await CreateSlugFromGroupNameAsync(groupName)
             };
 
             _db.Groups.Add(group);
             _db.Memberships.Add(new Membership
             {
                 Group = group, 
-                UserId = ownerId, 
+                UserId = CurrentUserId, 
                 DisplayName = ownerDisplayName, 
                 Role = GroupRole.Owner
             });
@@ -53,13 +59,15 @@ namespace LetsGame.Web.Services
             return group;
         }
 
-        public async Task<GroupGame> AddGameToGroupAsync(long groupId, long igdbId)
+        public async Task AddGameToGroupAsync(long groupId, long igdbId)
         {
+            await EnsureIsGroupOwnerAsync(groupId);
+
             var groupGame = await _db.GroupGames
                 .FirstOrDefaultAsync(x => x.GroupId == groupId &&
                                           x.IgdbId == igdbId);
             
-            if (groupGame != null) return groupGame;
+            if (groupGame != null) return;
             
             var game = await _igdbClient.GetGameAsync(igdbId);
             if (game == null) throw new ArgumentException("Game not found", nameof(igdbId));
@@ -74,26 +82,25 @@ namespace LetsGame.Web.Services
 
             _db.GroupGames.Add(groupGame);
             await _db.SaveChangesAsync();
-
-            return groupGame;
         }
 
-        public async Task AddSlotVoteAsync(long slotId, string voterId)
+        public async Task AddSlotVoteAsync(long slotId)
         {
             var slot = await _db.GroupEventSlots
                 .Include(x => x.Votes)
+                .Where(x => x.Event.Group.Memberships.Any(m => m.UserId == CurrentUserId))
                 .FirstOrDefaultAsync(x => x.Id == slotId);
             
-            if (slot != null && !slot.Votes.Any(v => v.VoterId == voterId))
+            if (slot != null && !slot.Votes.Any(v => v.VoterId == CurrentUserId))
             {
                 _db.GroupEventSlotVotes.Add(new GroupEventSlotVote
                 {
                     Slot = slot,
-                    VoterId = voterId,
+                    VoterId = CurrentUserId,
                     VotedAtUtc = DateTime.UtcNow
                 });
                 
-                var cantPlay = await _db.GroupEventCantPlays.FirstOrDefaultAsync(x => x.EventId == slot.EventId && x.UserId == voterId);
+                var cantPlay = await _db.GroupEventCantPlays.FirstOrDefaultAsync(x => x.EventId == slot.EventId && x.UserId == CurrentUserId);
                 if (cantPlay != null)
                     _db.GroupEventCantPlays.Remove(cantPlay);
                 
@@ -101,10 +108,10 @@ namespace LetsGame.Web.Services
             }
         }
 
-        public async Task RemoveSlotVoteAsync(long slotId, string voterId)
+        public async Task RemoveSlotVoteAsync(long slotId)
         {
             var vote = await _db.GroupEventSlotVotes
-                .FirstOrDefaultAsync(x => x.SlotId == slotId && x.VoterId == voterId);
+                .FirstOrDefaultAsync(x => x.SlotId == slotId && x.VoterId == CurrentUserId);
 
             if (vote != null)
             {
@@ -113,15 +120,20 @@ namespace LetsGame.Web.Services
             }
         }
 
-        public async Task SetCantPlayAsync(long eventId, string userId)
+        public async Task SetCantPlayAsync(long eventId)
         {
-            var votes = await _db.GroupEventSlotVotes.Where(x => x.Slot.EventId == eventId && x.VoterId == userId).ToListAsync();
+            var isGroupMember = await _db.GroupEvents
+                .Where(e => e.Group.Memberships.Any(m => m.UserId == CurrentUserId))
+                .AnyAsync(e => e.Id == eventId);
+            if (!isGroupMember) throw new InvalidOperationException("Not group member");
+            
+            var votes = await _db.GroupEventSlotVotes.Where(x => x.Slot.EventId == eventId && x.VoterId == CurrentUserId).ToListAsync();
             if (votes.Any()) _db.GroupEventSlotVotes.RemoveRange(votes);
 
-            var exists = await _db.GroupEventCantPlays.AnyAsync(x => x.EventId == eventId && x.UserId == userId);
+            var exists = await _db.GroupEventCantPlays.AnyAsync(x => x.EventId == eventId && x.UserId == CurrentUserId);
             if (!exists)
             {
-                _db.GroupEventCantPlays.Add(new GroupEventCantPlay {EventId = eventId, UserId = userId});
+                _db.GroupEventCantPlays.Add(new GroupEventCantPlay {EventId = eventId, UserId = CurrentUserId});
             }
 
             await _db.SaveChangesAsync();
@@ -131,14 +143,17 @@ namespace LetsGame.Web.Services
         {
             var slot = await _db.GroupEventSlots
                 .Include(x => x.Event)
+                .Where(x => x.Event.CreatorId == CurrentUserId || x.Event.Group.Memberships.Any(m => m.Role == GroupRole.Owner && m.UserId == CurrentUserId))
                 .FirstOrDefaultAsync(x => x.Id == slotId);
 
             slot.Event.ChosenDateAndTimeUtc = slot.ProposedDateAndTimeUtc;
             await _db.SaveChangesAsync();
         }
 
-        public async Task<GroupInvite> CreateInviteAsync(long groupId, bool singleUse)
+        public async Task CreateInviteAsync(long groupId, bool singleUse)
         {
+            await EnsureIsGroupOwnerAsync(groupId);
+            
             string id;
             
             do
@@ -155,25 +170,31 @@ namespace LetsGame.Web.Services
 
             _db.GroupInvites.Add(invite);
             await _db.SaveChangesAsync();
-
-            return invite;
         }
 
         public async Task DeleteInviteAsync(string id)
         {
-            var invite = await _db.GroupInvites.FindAsync(id);
+            var invite = await _db.GroupInvites
+                .Where(x => x.Group.Memberships.Any(m => m.Role == GroupRole.Owner && m.UserId == CurrentUserId))
+                .FirstOrDefaultAsync(x => x.Id == id);
+            
             if (invite != null) _db.GroupInvites.Remove(invite);
             
             await _db.SaveChangesAsync();
         }
 
-        public async Task<Group> AcceptInviteAsync(string userId, string displayName, string inviteId)
+        public async Task<Group> AcceptInviteAsync(string displayName, string inviteId)
         {
-            var invite = await _db.GroupInvites.Include(x => x.Group).FirstOrDefaultAsync(x => x.Id == inviteId);
+            var invite = await _db.GroupInvites
+                .Include(x => x.Group).ThenInclude(x => x.Memberships)
+                .FirstOrDefaultAsync(x => x.Id == inviteId);
+
+            if (invite.Group.Memberships.Any(m => m.UserId == CurrentUserId))
+                return invite.Group;
 
             _db.Memberships.Add(new Membership
             {
-                UserId = userId,
+                UserId = CurrentUserId,
                 GroupId = invite.GroupId,
                 DisplayName = displayName,
                 Role = GroupRole.Member
@@ -187,6 +208,24 @@ namespace LetsGame.Web.Services
             await _db.SaveChangesAsync();
 
             return invite.Group;
+        }
+
+        private string CurrentUserId => _userManager.GetUserId(_currentUserAccessor.CurrentUser);
+
+        private Task<string> CreateSlugFromGroupNameAsync(string name)
+        {
+            return _slugGenerator.GenerateWithCheck(name, slug => _db.Groups.AnyAsync(x => x.Slug == slug));
+        }
+
+        private async Task EnsureIsGroupOwnerAsync(long groupId)
+        {
+            var isGroupOwner = await _db.Groups
+                .AnyAsync(x => x.Id == groupId &&
+                               x.Memberships.Any(m =>
+                                   m.Role == GroupRole.Owner &&
+                                   m.UserId == CurrentUserId));
+            
+            if (!isGroupOwner) throw new InvalidOperationException("Not group owner");
         }
     }
 }
